micro-changes from JSC
svn path=/trunk/KDE/kdelibs/; revision=649649

+++ kjs/identifier.cpp
@@ -124,12 +124,12 @@ struct CStringTranslator
 
     static void translate(UString::Rep*& location, const char *c, unsigned hash)
     {
-        int length = strlen(c);
+        size_t length = strlen(c);
         UChar *d = static_cast<UChar *>(fastMalloc(sizeof(UChar) * length));
-        for (int i = 0; i != length; i++)
+        for (size_t i = 0; i != length; i++)
             d[i] = c[i];
 
-        UString::Rep *r = UString::Rep::create(d, length).release();
+        UString::Rep *r = UString::Rep::create(d, static_cast<int>(length)).release();
         r->isIdentifier = 1;
         r->rc = 0;
         r->_hash = hash;
@@ -142,7 +142,7 @@ PassRefPtr<UString::Rep> Identifier::add(const char *c)
 {
     if (!c)
         return &UString::Rep::null;
-    int length = strlen(c);
+    size_t length = strlen(c);
     if (length == 0)
         return &UString::Rep::empty;
